+++ tests/models/source_pipe_sink/src/double_pipe_heat.py
@@ -4,7 +4,9 @@ from mesido.esdl.esdl_parser import ESDLFileParser
 from mesido.esdl.profile_parser import ProfileReaderFromFile
 from mesido.techno_economic_mixin import TechnoEconomicMixin
 from mesido.workflows.io.write_output import ScenarioOutput
-
+from mesido.workflows.utils.adapt_profiles import (
+    adapt_profile_to_averaged_timestep,
+)
 
 from rtctools.optimization.collocated_integrated_optimization_problem import (
     CollocatedIntegratedOptimizationProblem,
@@ -73,6 +75,23 @@ class SourcePipeSink(
 
     def post(self):
         super().post()
+
+
+class SourcePipeSinkDayAveraged(SourcePipeSink):
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__day_steps = kwargs.get("_day_steps", 1)
+
+    def parameters(self, ensemble_member):
+        parameters = super().parameters(ensemble_member)
+        parameters["time_step_days"] = self.__day_steps
+        return parameters
+
+    def read(self):
+        super().read()
+
+        adapt_profile_to_averaged_timestep(self, self.__day_steps * 24)
 
 
 class HeatProblemESDLVarsMixin(ESDLAdditionalVarsMixin, SourcePipeSink):