+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# [Unreleased-main] - 2026-07-14
+# [Unreleased-main] - 2026-07-15
 
 ## Added
 - ESDL profile reading and writing for different types: influxdb and postgres, and inline (in ESDL file).
+- Parsing of input profiles with 5-minute interval
 
 ## Changed
 - Code generalization in asset sizing mixin